Fallback activado en BDD, WIP de tarjetas en home
En teoria room va a resetear solo la memoria si no puede hacer
migraciones ahora. (En lugar de crashear en primer acceso)

Home ya muestra las tarjetas segun cada usuario, pero solo para
formulario 1 (Transectos). En la mañana voy a implementarlo para los
demás. Debería de ser sencillo. Ya implementé lo que podría ser
la pequeña capa de abtracción para las tarjetas. Esta en Home.kt

"Debería de ser sencillo"
	 - Omar de hoy

+++ app/src/main/java/com/example/awaq1/data/formularios/FormulariosDAO.kt
@@ -20,7 +20,7 @@ interface FormularioUnoDAO {
     suspend fun delete(item: FormularioUnoEntity)
 
     @Query("SELECT * from Formulario1 WHERE id = :id")
-    fun getFormularioUnoEntity(id: Int): Flow<FormularioUnoEntity?>
+    fun getFormularioUnoEntity(id: Long): Flow<FormularioUnoEntity?>
 
     @Query("SELECT * from Formulario1 ORDER BY id ASC")
     fun getAllFormularioUnoEntities(): Flow<List<FormularioUnoEntity>>
@@ -41,7 +41,7 @@ interface FormularioDosDAO {
     suspend fun delete(item: FormularioDosEntity)
 
     @Query("SELECT * from Formulario2 WHERE id = :id")
-    fun getFormularioDosEntity(id: Int): Flow<FormularioDosEntity?>
+    fun getFormularioDosEntity(id: Long): Flow<FormularioDosEntity?>
 
     @Query("SELECT * from Formulario2 ORDER BY id ASC")
     fun getAllFormularioDosEntities(): Flow<List<FormularioDosEntity>>
@@ -62,7 +62,7 @@ interface FormularioTresDAO {
     suspend fun delete(item: FormularioTresEntity)
 
     @Query("SELECT * from Formulario3 WHERE id = :id")
-    fun getFormularioTresEntity(id: Int): Flow<FormularioTresEntity?>
+    fun getFormularioTresEntity(id: Long): Flow<FormularioTresEntity?>
 
     @Query("SELECT * from Formulario3 ORDER BY id ASC")
     fun getAllFormularioTresEntities(): Flow<List<FormularioTresEntity>>
@@ -83,7 +83,7 @@ interface FormularioCuatroDAO {
     suspend fun delete(item: FormularioCuatroEntity)
 
     @Query("SELECT * from Formulario4 WHERE id = :id")
-    fun getFormularioCuatroEntity(id: Int): Flow<FormularioCuatroEntity?>
+    fun getFormularioCuatroEntity(id: Long): Flow<FormularioCuatroEntity?>
 
     @Query("SELECT * from Formulario4 ORDER BY id ASC")
     fun getAllFormularioCuatroEntities(): Flow<List<FormularioCuatroEntity>>
